django/pyproject.toml
# Adapted from https://github.com/Grvs44/budgetmanager/blob/main/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "inclusivevenues"
version = "1.0.1"
authors = [
    { name="Elli Greaves" },
]
license = { text = "BSD-3-Clause-Clear license" }
requires-python = ">=3.10,<4"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
]
dependencies = [
    "django>=5.1.5",
    "djangorestframework>=3.15.2",
    "django-filter>=24.3",
    "pillow>=11.1.0",
    "requests>=2.32.3"
]

[project.urls]
"Homepage" = "https://github.com/Grvs44/inclusivevenues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["inclusivevenues", "inclusivevenues.*"]
